% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockwork.R
\name{clockwork}
\alias{clockwork}
\title{Create a clockwork widget}
\usage{
clockwork(data, x, y, cycle, x_ticks = 4, y_ticks = 3,
  show_cycle_stats = FALSE, cycle_label = "Cycle: ",
  width_height = 400, loop_interval = 4000, repeat_cycles = TRUE)
}
\arguments{
\item{data}{a tabular object}

\item{x}{a numeric variable}

\item{y}{a numeric variable}

\item{cycle}{a variable to cycle/overlay subsequent lines}

\item{x_ticks}{number of ticks for x-axis}

\item{y_ticks}{number of ticks for y-axis}

\item{show_cycle_stats}{whether to display the min, max, and mean for each cycle}

\item{cycle_label}{cycle label}

\item{width_height}{widget width and height}

\item{loop_interval}{cycle length in milliseconds}

\item{repeat_cycles}{whether to repeat visualization after all cycles}
}
\description{
Create a clockwork widget
}
