% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockwork.R
\name{cw_style}
\alias{cw_style}
\title{Style a clockwork widget}
\usage{
cw_style(clockwork, active_line_color = "#da4f81",
  past_line_color = "#ccc", past_line_opacity = 0.4,
  grid_line_color = "#000", x_axis_font_size = 10,
  y_axis_font_size = 10, inner_stats_font_size = 12,
  font_family = "Arial")
}
\arguments{
\item{clockwork}{a clockwork object}

\item{active_line_color}{color of active line being drawn}

\item{past_line_color}{color of past lines drawn}

\item{past_line_opacity}{opacity of past lines drawn}

\item{grid_line_color}{color of grid lines}

\item{x_axis_font_size}{x-axis label font-size}

\item{y_axis_font_size}{y-axis label font-size}

\item{inner_stats_font_size}{inner stats font size}

\item{font_family}{font family}
}
\value{
a clockwork widget
}
\description{
Style a clockwork widget
}
